4.2 Работа на проверку (исправление)

+++ scripts422.sql
@@ -2,7 +2,7 @@ CREATE TABLE person
 (
     id SERIAL PRIMARY KEY,
     name TEXT,
-    age SERIAL,
+    age INTEGER,
     driver_id BOOLEAN,
     car_id SERIAL REFERENCES car (id)
 )
@@ -12,5 +12,5 @@ CREATE TABLE car
 id SERIAL PRIMARY KEY,
 brand TEXT,
 model TEXT,
-price SERIAL
+price REAL
 )
